+++ api/custom_auth/sso/oauth/github.py
@@ -4,8 +4,8 @@ import requests
 from django.conf import settings
 from requests import RequestException
 
-from custom_auth.oauth.exceptions import GithubError
-from custom_auth.oauth.helpers.github_helpers import (
+from custom_auth.sso.oauth.exceptions import GithubError
+from custom_auth.sso.oauth.helpers.github_helpers import (
     convert_response_data_to_dictionary,
     get_first_and_last_name,
 )
