+++ api/custom_auth/sso/oauth/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from custom_auth.oauth.views import login_with_github, login_with_google
+from custom_auth.sso.oauth.views import login_with_github, login_with_google
 
 app_name = "oauth"
