+++ api/custom_auth/urls.py
@@ -23,5 +23,5 @@ urlpatterns = [
     path("", include("djoser.urls")),
     path("", include("trench.urls")),  # MFA
     path("", include("trench.urls.djoser")),  # override necessary urls for MFA auth
-    path("oauth/", include("custom_auth.oauth.urls")),
+    path("oauth/", include("custom_auth.sso.oauth.urls")),
 ]
